getStationReport service is implemented

+++ booking-service/src/main/java/com/gearup/model/StationReport.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class StationReport {
     private  Long stationId;
     private  String stationName;
-    private  Double totalEarnings;
+    private  int totalEarnings;
     private  Integer totalBookings;
     private  Integer cancelledBooking;
     private  Double totalRefund;
